Refactor and simplify
This commit removes a lot of classes and complexity. I did this because
I think it's helpful to keep the Scala/Java boundary as small as
possible in order to avoid things like casts from `Any`. Instead, the
default SnakeYAML classes are used for serialization.

The interface is now much more inline with the original interface (and
with circe itself). All of the SnakeYAML details are hidden from the
functional interface, so another implementation could implement the same
package structure and functional interface without needing any
inheritance.

The symmetric serialization tests pass; it still might be helpful to add
more tests, though.