Work-around for buggy treatment of 0x85 and 0xFEFF.
This fixes buggy encodings of strings, at the cost of a single pass
over the entire document to be rendered. Basically, if we detect a
string containing either of these characters, we fall back to a more
conservative set of SnakeYAML options.

EscaptingTests should ensure that at the character level our escaping
is correct. It also has a semi-duplicative test around round-tripping
strings, which also ensures that we only generate "printable
characters" when rendering raw YAML.