Remove myself from maintainers
Honestly, I haven't been able to use any recent version of `circe-yaml` (though I am still using an old version of it) because lately Spark has the majority of my projects stuck on Scala 2.11 (2.12 in some cases, but only in cases where it also has to work on 2.11, and that's understandably no longer supported). I guess my OSS badness is showing, but it's hard to find the time to actively maintain things that I'm unable to use (regrettably this is far from the only example 😞 ).

So I think it's only fair that I remove myself from the list of maintainers – even though I started `circe-yaml`, I haven't contributed to its maintenance in years (apologies to @travisbrown for transferring this thing to the circe org and then eventually abandoning it, foisting maintenance onto you and @jeffmay) so I don't think I deserve any credit as a maintainer now.

I also edited to credit @travisbrown as a maintainer, given that (from what the GitHub history tells me) he's the primary person who's actively doing maintenance (apologies again!). I'm unsure whether @jeffmay is in the same boat as me, but surely @travisbrown seems to be doing the bulk of the legwork recently, so I'll ping @jeffmay one more time to provoke them to consider it 😄